pyproject.toml
[project]
name = "homelab-brain"
version = "0.1.0"
description = "Natural language interface for homelab management with AI (Winston for your homelab!)"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "ollama>=0.4",
    "psutil>=5.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
]

[project.scripts]
lab = "homelab_brain.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["homelab_brain*", "config*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
